Fixed bug affecting iCloud Drive paths
File paths in iCloud Drive tend to contain tildes. The `expandPath` function
was greedily replacing these instead of just replacing the leading `~/` home path
shortcut as intended.

The function now uses a regexp so that only the tilde at the start of
the path (if it exists) is replaced with the expanded home path. Any
other tildes are left in peace.

Fixes #1